Move inlining to it's own phase
